requirements.txt
discord.py == 1.4.1
tweepy
parsedatetime
python-dateutil
dispander
youtube_dl
pynacl